requirements.txt
annotated-doc==0.0.4
anyio==4.12.1
audioop-lts==0.2.2
av==16.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comtypes==1.4.15
ctranslate2==4.7.1
faster-whisper==1.2.1
filelock==3.24.3
flatbuffers==25.12.19
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.5.0
idna==3.11
Jinja2==3.1.6
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==10.8.0
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.2
onnxruntime==1.24.2
packaging==26.0
pillow==12.0.0
protobuf==7.34.0
psutil==7.2.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycaw==20251023
pycparser==3.0
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
pyperclip==1.11.0
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.18.0
PyRect==0.2.0
PyScreeze==1.0.1
pyttsx3==2.99
pytweening==1.2.0
pywin32==311
PyYAML==6.0.3
regex==2026.2.19
requests==2.32.5
rich==14.3.3
screen_brightness_control==0.27.0
setuptools==82.0.0
shellingham==1.5.4
soundfile==0.13.1
SpeechRecognition==3.14.5
standard-aifc==3.13.0
standard-chunk==3.13.0
sympy==1.14.0
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typing_extensions==4.15.0
urllib3==2.6.3
webrtcvad-wheels==2.0.14
WMI==1.5.1
